requirements.txt
# Test dependencies
coverage>=7.4.0  # For test coverage reporting
pytest>=8.0.0    # Additional testing framework support

# Data processing and analysis
pandas>=2.2.0    # For data manipulation and analysis
numpy>=1.26.0    # For numerical computations
yfinance>=0.2.35 # For downloading financial data
ta>=0.10.0       # For technical analysis indicators

# Visualization
matplotlib>=3.8.0  # For plotting and visualization
seaborn>=0.13.0   # For enhanced visualizations

# Machine Learning
scikit-learn>=1.4.0  # For ML models and grid search

# Risk Management
scipy>=1.12.0  # For statistical calculations

# Utilities
python-dotenv>=1.0.0  # For environment variable management
requests>=2.31.0      # For HTTP requests
tqdm>=4.66.0         # For progress bars
PyYAML>=6.0.1        # For configuration management

# Development tools
black>=24.1.0     # Code formatting
flake8>=7.0.0     # Code linting
isort>=5.13.0     # Import sorting 